+++ exec/demo_empty.sh
@@ -5,7 +5,10 @@
 #
 # Special case: all empty input files
 
-dir=$(readlink -f $(dirname $BASH_SOURCE))
+cd "$(dirname $BASH_SOURCE)"
+dir=$(pwd -P)
+inst="../inst"
+[ -d "$inst" ] || inst=".."
 scratch=$(mktemp -d)
 mkdir -p "$scratch/str-dataset"
 for samp in {1..3}; do
@@ -13,7 +16,8 @@ for samp in {1..3}; do
 		touch "$scratch/str-dataset/Replicate1-Sample${samp}-${locus}.fasta"
 	done
 done
-cp "$dir/../example_locus_attrs.csv" "$scratch/locus_attrs.csv"
-cp "$dir/../example_config.yml" "$scratch/config.yml"
+cp "$dir/$inst/example_locus_attrs.csv" "$scratch/locus_attrs.csv"
+cp "$dir/$inst/example_config.yml" "$scratch/config.yml"
 cd "$scratch"
 "$dir/chiimp" "config.yml"
+echo "Demo files and output stored in $(pwd -P)"